backend/requirements.txt
fastapi
uvicorn
firebase-admin
sqlmodel
sagemaker
python-dotenv
